Refactoring against 31.2x64
