+++ Nodes/VVVV.Time.Nodes/Core/DateTimeWithZone.cs
@@ -1,9 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+using System;
 
-namespace VVVV.Time
+namespace Time.Core
 {
     public struct DateTimeWithZone
     {
