+++ Nodes/VVVV.Time.Nodes/Core/SortEnum.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace VVVV.Packs.Time
+{
+    public enum SortEnum
+    {
+        Ascending,
+        Descending
+    }
+}
