+++ Nodes/VVVV.Time.Nodes/VVVV.Time.Nodes.sln
@@ -1,0 +1,36 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 11.00
+# Visual Studio 2010
+# SharpDevelop 4.1.0.8000
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VVVV.Time.Nodes", "VVVV.Time.Nodes.csproj", "{F8924570-16ED-4425-A64C-17CC05A41380}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F8924570-16ED-4425-A64C-17CC05A41380}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F8924570-16ED-4425-A64C-17CC05A41380}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F8924570-16ED-4425-A64C-17CC05A41380}.Debug|x64.Build.0 = Debug|x64
+		{F8924570-16ED-4425-A64C-17CC05A41380}.Debug|x64.ActiveCfg = Debug|x64
+		{F8924570-16ED-4425-A64C-17CC05A41380}.Debug|x86.Build.0 = Debug|x86
+		{F8924570-16ED-4425-A64C-17CC05A41380}.Debug|x86.ActiveCfg = Debug|x86
+		{F8924570-16ED-4425-A64C-17CC05A41380}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F8924570-16ED-4425-A64C-17CC05A41380}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F8924570-16ED-4425-A64C-17CC05A41380}.Release|x64.Build.0 = Release|x64
+		{F8924570-16ED-4425-A64C-17CC05A41380}.Release|x64.ActiveCfg = Release|x64
+		{F8924570-16ED-4425-A64C-17CC05A41380}.Release|x86.Build.0 = Release|x86
+		{F8924570-16ED-4425-A64C-17CC05A41380}.Release|x86.ActiveCfg = Release|x86
+	EndGlobalSection
+EndGlobal
