+++ vvvv-Time.sln
@@ -15,8 +15,8 @@ Global
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{F8924570-16ED-4425-A64C-17CC05A41380}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F8924570-16ED-4425-A64C-17CC05A41380}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F8924570-16ED-4425-A64C-17CC05A41380}.Debug|x64.ActiveCfg = Debug|x86
-		{F8924570-16ED-4425-A64C-17CC05A41380}.Debug|x64.Build.0 = Debug|x86
+		{F8924570-16ED-4425-A64C-17CC05A41380}.Debug|x64.ActiveCfg = Debug|x64
+		{F8924570-16ED-4425-A64C-17CC05A41380}.Debug|x64.Build.0 = Debug|x64
 		{F8924570-16ED-4425-A64C-17CC05A41380}.Debug|x86.ActiveCfg = Debug|x86
 		{F8924570-16ED-4425-A64C-17CC05A41380}.Debug|x86.Build.0 = Debug|x86
 		{F8924570-16ED-4425-A64C-17CC05A41380}.Release|Any CPU.ActiveCfg = Release|Any CPU
